refactor(seederTest): refactor seeder test
