+++ test/unit/Seeder.spec.ts
@@ -4,41 +4,10 @@ import {Wetland} from '../../src/Wetland';
 import * as path from 'path';
 import * as Bluebird from 'bluebird';
 import * as fs from 'fs';
-import * as rimraf from 'rimraf';
-import * as parse  from 'csv-parse';
+import * as parse from 'csv-parse';
+import {rmDataDir, tmpTestDir, getType, Pet, Post, User, fixturesDir} from '../resource/Seeder';
 
-const tmpTestDir  = path.join(__dirname, '../.tmp');
-const dataDir     = `${tmpTestDir}/.data`;
-const fixturesDir = path.join(__dirname, '../resource/fixtures/');
-
-class User {
-  static setMapping(mapping) {
-    mapping.forProperty('id').increments().primary();
-    mapping.forProperty('username').field({type: 'string'});
-    mapping.forProperty('password').field({type: 'string'});
-    mapping.forProperty('posts').oneToMany({targetEntity: Post, mappedBy: 'author'});
-  }
-}
-
-class Post {
-  static setMapping(mapping) {
-    mapping.forProperty('id').increments().primary();
-    mapping.forProperty('title').field({type: 'string'});
-    mapping.forProperty('content').field({type: 'text'});
-    mapping.forProperty('author').manyToOne({targetEntity: User, inversedBy: 'posts'});
-  }
-}
-
-class Pet {
-  static setMapping(mapping) {
-    mapping.forProperty('id').increments().primary();
-    mapping.forProperty('name').field({type: 'string'});
-  }
-}
-
-const getType = (bypassLifecyclehooks: boolean) => bypassLifecyclehooks ? 'nolifecycle' : 'lifecycle';
-
-const getWetland = (clean: boolean, bypassLifecyclehooks: boolean) => {
+function getWetland(clean: boolean, bypassLifecyclehooks: boolean): Wetland {
   const fileName = `${clean ? 'clean' : 'safe'}-${getType(bypassLifecyclehooks)}.sqlite`;
   return new Wetland({
     dataDirectory: `${tmpTestDir}/.data`,
@@ -58,9 +27,9 @@ const getWetland = (clean: boolean, bypassLifecyclehooks: boolean) => {
     },
     entities     : [User, Pet, Post]
   });
-};
+}
 
-const testUsers = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean) => {
+function testUsers(manager: Scope, bypassLifecyclehooks: boolean): Promise<any> {
   let usersFromFile = require(path.join(fixturesDir, getType(bypassLifecyclehooks), 'User.json'));
 
   return manager.getRepository(User)
@@ -74,9 +43,9 @@ const testUsers = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean
         });
       }
     });
-};
+}
 
-const testPosts = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean) => {
+function testPosts(manager: Scope, bypassLifecyclehooks: boolean): Promise<any> {
   let postsFromFile = require(path.join(fixturesDir, getType(bypassLifecyclehooks), 'Post.json'));
 
   return manager.getRepository(Post)
@@ -90,9 +59,9 @@ const testPosts = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean
         });
       }
     });
-};
+}
 
-const testPets = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean) => {
+function testPets(manager: Scope, bypassLifecyclehooks: boolean): Promise<any> {
   const readFile: any = Bluebird.promisify(fs.readFile);
 
   return readFile(path.join(fixturesDir, getType(bypassLifecyclehooks), 'Pet.csv'))
@@ -107,14 +76,10 @@ const testPets = (manager: Scope, clean: boolean, bypassLifecyclehooks: boolean)
           assert.lengthOf(pets, petsFromFile.length);
         });
     });
-};
+}
 
 describe('Seeder', () => {
-  beforeEach(() => {
-    const rmDir: any = Bluebird.promisify(rimraf);
-
-    return rmDir(dataDir)
-  });
+  beforeEach(() => rmDataDir());
 
   describe('CleanSeed', () => {
     describe('.seed(): no lifecyclehooks', () => {
@@ -125,14 +90,15 @@ describe('Seeder', () => {
         const wetland  = getWetland(clean, bypassLifecyclehooks);
         const migrator = wetland.getMigrator();
         const seeder   = wetland.getSeeder();
+        const cleaner  = wetland.getCleaner();
         const manager  = wetland.getManager();
 
-        return seeder.clean()
+        return cleaner.clean()
           .then(() => migrator.devMigrations(false))
           .then(() => seeder.seed())
-          .then(() => testUsers(manager, clean, bypassLifecyclehooks))
-          .then(() => testPosts(manager, clean, bypassLifecyclehooks))
-          .then(() => testPets(manager, clean, bypassLifecyclehooks));
+          .then(() => testUsers(manager, bypassLifecyclehooks))
+          .then(() => testPosts(manager, bypassLifecyclehooks))
+          .then(() => testPets(manager, bypassLifecyclehooks));
       });
     });
 
@@ -145,13 +111,14 @@ describe('Seeder', () => {
         const migrator = wetland.getMigrator();
         const seeder   = wetland.getSeeder();
         const manager  = wetland.getManager();
+        const cleaner  = wetland.getCleaner();
 
-        return seeder.clean()
+        return cleaner.clean()
           .then(() => migrator.devMigrations(false))
           .then(() => seeder.seed())
-          .then(() => testUsers(manager, clean, bypassLifecyclehooks))
-          .then(() => testPosts(manager, clean, bypassLifecyclehooks))
-          .then(() => testPets(manager, clean, bypassLifecyclehooks));
+          .then(() => testUsers(manager, bypassLifecyclehooks))
+          .then(() => testPosts(manager, bypassLifecyclehooks))
+          .then(() => testPets(manager, bypassLifecyclehooks));
       });
     });
   });
@@ -170,9 +137,9 @@ describe('Seeder', () => {
         return migrator.devMigrations(false)
           .then(() => seeder.seed())
           .then(() => seeder.seed()) // Called a second time on purpose
-          .then(() => testUsers(manager, clean, bypassLifecyclehooks))
-          .then(() => testPosts(manager, clean, bypassLifecyclehooks))
-          .then(() => testPets(manager, clean, bypassLifecyclehooks));
+          .then(() => testUsers(manager, bypassLifecyclehooks))
+          .then(() => testPosts(manager, bypassLifecyclehooks))
+          .then(() => testPets(manager, bypassLifecyclehooks));
       });
     });
 
@@ -189,9 +156,9 @@ describe('Seeder', () => {
         return migrator.devMigrations(false)
           .then(() => seeder.seed())
           .then(() => seeder.seed()) // Called a second time on purpose
-          .then(() => testUsers(manager, clean, bypassLifecyclehooks))
-          .then(() => testPosts(manager, clean, bypassLifecyclehooks))
-          .then(() => testPets(manager, clean, bypassLifecyclehooks));
+          .then(() => testUsers(manager, bypassLifecyclehooks))
+          .then(() => testPosts(manager, bypassLifecyclehooks))
+          .then(() => testPets(manager, bypassLifecyclehooks));
       });
     });
   });
